+++ src/main/java/it/sirziphonprojects/hibernateentitymodel/entity/KeyMapper.java
@@ -3,10 +3,10 @@ package it.sirziphonprojects.hibernateentitymodel.entity;
 import java.util.Map;
 
 /**
- * Interface for key objects of entities
+ * This interface provides a set of methods for managing entity keys
  *
  * @author SirZiphon {@literal <https://github.com/sirziphon>}
- * @version 1.0.0
+ * @version 1.1.0
  */
 public interface KeyMapper {
 
@@ -24,5 +24,39 @@ public interface KeyMapper {
      * @param keysMap
      * @return
      */
-    public boolean equals(Map<String, Object> keysMap);
+    public default boolean equals(Map<String, Object> keysMap) {
+        // check if keysMap is null or empty
+        if(keysMap == null || keysMap.isEmpty())
+            return false;
+
+        // get local keysMap
+        Map<String, Object> localMap = this.getAsMap();
+
+        // check if the keysMap have the same number of keys
+        if (localMap.size() != keysMap.size())
+            return false;
+
+        // confront any camp
+        for (String key : localMap.keySet()) {
+            if (!localMap.get(key).equals(keysMap.get(key)))
+                return false;
+        }
+
+        return true;
+    }
+
+    /**
+     * this method confront the current instance with the one passed as parameter; return true if the keys match,
+     * <code>false</code> otherwise
+     *
+     * @param keyMapper
+     * @return
+     */
+    public default boolean equals(KeyMapper keyMapper) {
+        // check if keyMapper is null
+        if (keyMapper == null)
+            return false;
+
+        return this.equals(keyMapper.getAsMap());
+    }
 }